+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Pinguino Changelog
 
+## Version 0.7.8
+
+**[Tag Comparison](https://github.com/JamCoreDiscord/Pinguino/compare/v0.7.7...v0.7.8)
+
+### Fixes
+
+### Features
+
 ## Version 0.7.7
 
 **[Tag Comparison](https://github.com/JamCoreDiscord/Pinguino/compare/v0.7.6...v0.7.7)**